modify Hits(unsigned int plane, unsigned int wire=-1) to be Hits(unsigned int plane, unsigned int wire=0, bool allwires=true)
so now the default behavior of the method is to return all hits on 
the specified plane.  for example

edm::PtrVector<recob::Hits> clusthits = mycluster->Hits(p);

will return all hits from the edm::Ptr<recob::Cluster> defined as mycluster 
for plane numbered p.

you get specific hits from a given wire on the plane by setting the wire 
and allwires arguments, ie

edm::PtrVector<recob::Hits> clusthits = mycluster->Hits(p,w,false);

this is a little clunky, as one would prefer if the wire is set then that 
defines that only a specific wire is wanted, however, there is no good default
value for the wire argument that works as each TPC has a different number of
wires on each plane.  -1 doesnt work as wire has to be an unsigned int.

with this change all calls in ClusterFinder modules like 

mycluster->Hits(p,-1);

should now be 

mycluster->Hits(p);

fyi, the -1 was never needed before anyway as that was the default argument in
the Cluster.h definition of the interface.




Former-commit-id: 8e1c4095efd42b85240083814c89ae9bc8c80f3d