Initial check-in of SpacePoints class.  SpacePoints will store 3-dimensional information coming from Hits that have been matched from multiple views.  This will allow us to avoid creating a new Hits collections for 3d info, which will reduce bloat.
Have started to update other RecoBase classes (most notably Prong/Track/Shower) to reflect this change.  The constructor for these classes now can also take a std::vector of SpacePoints (the std::vector is default initialized so as not to break existing algorithms like ShowerFinder that don't need to create SpacePoint collections).  

To-do:  Need to remove the XYZ info. from the Hits class, and then propogate these changes to the relevant reconstruction algorithms. 

- Mitch




Former-commit-id: 39c8e3d827e33b0c2ce3b39b8b452105de1503df