Update of original Track3Dreco code to reflect changes in RecoBase.  Biggest change to RecoBase was the removal of the XYZ info. from the Hit class, and the creation of the SpacePoint class to hold this information.  Track3Dreco now reflects those changes and compiles, but I have not had an opportunity to check that is doing something sane when it runs.  Hence, don't use this yet unless you're just curious.
Still need to rework this code further to be more detector agnostic, and to possibly for any pair of planes, and not just for the last Induction and Collection plane.   

The GNUmakefile also is updated to work for the TrackFinder class...it must have been copied from the ClusterFinder package initialy, and had the wrong namespace info.

-Mitch




Former-commit-id: f3e58045cba142bf1bf9aa1f2e2f226f0103e7f3