Initial import of AnalysisExample package.
The target audience for these are files are users who are brand-new to
LArSoft, and whose goal is to analyze data in LArSoft event files. It
demonstrates how to read art::Event records and create histograms and
n-tuples. There are comments everywhere to help the LArSoft new-comer
understand why things work.

This package goes along with a wiki page:
<https://cdcvs.fnal.gov/redmine/projects/larsoftsvn/wiki/Getting_started_with_an_analysis_task>




Former-commit-id: 3b500d497b881b9ff0eba98498db380292ef9fd8